chore(): add versions dropdown to custom toolbar
