+++ packages/react-ui/.storybook/Meta.tsx
@@ -15,8 +15,9 @@ import { reactUIFeatureFlagsDefault } from '../lib/featureFlagsContext';
 import { MenuSeparator } from '../components/MenuSeparator';
 import { MenuFooter } from '../components/MenuFooter';
 import { Link } from '../components/Link';
+import { VersionsLibrary } from './VersionsDropdown/Versions';
 
-const urlPath = window.location.origin + window.location.pathname;
+const urlPath = window.location.origin + window.location.pathname.replace('/iframe.html', '');
 
 const languages = [
   { icon: '🇷🇺', caption: 'Russian', value: 'ru' },
@@ -221,6 +222,16 @@ export const Meta = ({ of }: { of: ModuleExports }) => {
               Подробнее о фича-флагах
             </Link>
           </MenuFooter>
+        </DropdownMenu>
+
+        <DropdownMenu
+          style={{ marginLeft: 'auto', marginRight: '16px' }}
+          caption={<div className={styles.menuSelect}>v {urlPath.split('/').pop()?.replace('-', ' ')}</div>}
+          menuMaxHeight="315px"
+          menuWidth="150px"
+          positions={['bottom right']}
+        >
+          <VersionsLibrary />
         </DropdownMenu>
       </div>
     </div>
