--- packages/react-ui/.storybook/docsPages/Versioning/Versions.mdx
@@ -1,10 +1,0 @@
-import { VersionsLibrary } from './Versions';
-import { Meta, Unstyled } from '@storybook/blocks';
-
-<Meta title="Versioning/Versions" />
-
-<Unstyled id="disable-toc">
-  # Версии React UI
-
-  <VersionsLibrary />
-</Unstyled>
